+++ packages/hooks/src/createDidHook.js
@@ -1,0 +1,17 @@
+/**
+ * Retrieve the number of times an action is fired.
+ *
+ * @param {string} hooksArray Hooks array of hooks to check.
+ * @param {string} action     The action to check.
+ *
+ * @return {number}           The number of times the hook has run.
+ */
+const createDidHook = function( hooksArray ) {
+	return function( action ) {
+		return hooksArray && hooksArray[ action ] && hooksArray[ action ].runs ?
+			hooksArray[ action ].runs :
+			0;
+	};
+}
+
+export default createDidHook;
